CryptoPriceAlerts/requirements.txt
Django==3.2.16
django-extensions==3.2.1
djangorestframework==3.14.0
ipdb==0.13.13
ipython==8.12.0
dj-database-url==1.3.0
psycopg2==2.9.3
celery==5.2.7
redis==4.5.4
django-redis==5.2.0
djangorestframework-simplejwt==5.2.2
requests==2.29.0